+ modfiy RandomizedAreaObject health from int to float + mofify Damageable interface damage method to accept float parameter + remove visual Area boundaries

+++ core/src/szm/orde4c/game/entity/Damageable.java
@@ -1,6 +1,6 @@
 package szm.orde4c.game.entity;
 
 public interface Damageable {
-    void damage(int damage);
+    void damage(float damage);
     boolean isDestroyed();
 }
